UserExperienceAPIs

+++ LibraryAPI/DTOs/MembershipDTO.cs
@@ -4,7 +4,7 @@
     {
         public string MembershipName { get; set; }
         public int MaxBorrowLimit { get; set; }
-        public float MembershipPrice { get; set; }
+        public int MembershipPrice { get; set; }
         public int User_UserID { get; set; }
     }
 }
